Use SciMLTesting v1.2 folder-based run_tests
Replace the hand-written GROUP dispatch in test/runtests.jl with the
SciMLTesting v1.2 folder-discovery harness (`using SciMLTesting; run_tests()`).

Core = the top-level test/*.jl files (each run isolated under @safetestset by
the harness), QA = test/qa/ (sub-env activated automatically). The Core files
are wrapped in @safetestset for module isolation (each carries its own
`using NBodySimulator, ...`); show/string assertions keep printing unqualified
type names because the active module stays Main inside a @safetestset.

custom_potential_test.jl's body (which extends
NBodySimulator.get_accelerating_function via a module-top-level `import`) lives
in test/shared/custom_potential_body.jl and is included explicitly; test/shared/
is not a declared group, so it is not auto-discovered as Core.

Deps: add SciMLTesting + SafeTestsets to the root [extras]/test target/[compat]
and to test/qa/Project.toml; drop Pkg and its qa.jl activation preamble (the
harness owns all Pkg ops now). test/test_groups.toml unchanged. The exact set
of tests run under each GROUP value is preserved.
