+++ Project.toml
@@ -37,6 +37,8 @@ PrecompileTools = "1.2.0"
 RecipesBase = "1.3.4"
 RecursiveArrayTools = "3.36.0, 4"
 Reexport = "1.2.0"
+SafeTestsets = "0.1, 1"
+SciMLTesting = "1"
 StaticArrays = "1.9.18"
 StochasticDiffEq = "6.80.0, 7"
 Test = "1"
@@ -44,10 +46,11 @@ julia = "1.10"
 
 [extras]
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
+SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
+SciMLTesting = "09d9d899-5365-40a9-917a-5f67fddea283"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 StochasticDiffEq = "789caeaf-c7a9-5a7d-9973-96adeb23e2a0"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Pkg", "StaticArrays", "StochasticDiffEq", "LinearAlgebra", "Test"]
+test = ["LinearAlgebra", "SafeTestsets", "SciMLTesting", "StaticArrays", "StochasticDiffEq", "Test"]
