+++ test/runtests.jl
@@ -1,20 +1,2 @@
-using Test
-
-const GROUP = get(ENV, "GROUP", "All")
-
-if GROUP == "All" || GROUP == "Core"
-    using NBodySimulator, StaticArrays, LinearAlgebra, StochasticDiffEq
-
-    include("lennard_jones_test.jl")
-    include("electrostatics_test.jl")
-    include("gravitational_test.jl")
-    include("custom_potential_test.jl")
-    include("magnetostaic_test.jl")
-    include("thermostat_test.jl")
-    include("water_test.jl")
-    include("interface_test.jl")
-end
-
-if GROUP == "QA"
-    include("qa/qa.jl")
-end
+using SciMLTesting
+run_tests()
